pyproject.toml
[tool.poetry]
name = "fasthtml tutorial"
version = "0.1.0"
description = "FastHTML Tutorial Code"
authors = ["R. Bosschaert <rubbieus2@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.10.0"
matplotlib = "^3.9.2"
scipy = "^1.14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
